backend/pyproject.toml
[project]
name = "bmad-web-hub"
version = "0.1.0"
description = "Browser-based BMAD collaboration platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
